refactor: edit app styles, move app files to folder
