--- src/app.scss
@@ -1,10 +1,0 @@
-.app-wrapper {
-    display: flex;
-}
-
-.app-main {
-    display: flex;
-    flex: 1 90%;
-    justify-content: center;
-    background: #2C2F33
-}
