+++ src/components/checkbox/Checkbox.scss
@@ -1,41 +1,37 @@
-$element-size: 30px;
-$element-color: #25CEDE;
-$element-hover-cover: rgba(37, 206, 222, 0.5);
-$box-shadow-first-color: #000000;
-$box-shadow-second-color: #3B4451;
-$transition-duration: 0.2s;
+@import '../../styles/mixin.scss';
 
-@mixin size($width, $height: $width) {
-  width: $width;
-  height: $height;
-}
+@import '../../styles/variables.scss';
+
+$element-size: 30px;
+$transition-duration: 0.2s;
+$border-radius: 40%;
 
 .checkbox {
-  position: relative;
-  background-color:#2C2F33;
-  border-radius: 40%;
-  box-shadow: 6px 6px 18px $box-shadow-first-color, -6px -6px 18px $box-shadow-second-color;  
-  color: $element-color;
+  position: relative;  
+  color: $enabled-color;
+  background-color: $background-color;
+  border-radius: $border-radius;
+  box-shadow: $container-box-shadow;
+  outline: none;
+  transition: box-shadow $transition-duration, color $transition-duration;
+  cursor: pointer;
   -webkit-appearance: none;
   -moz-appearance: none;
   appearance: none;
-  outline: none;
-  cursor: pointer;
-  transition: box-shadow $transition-duration, color $transition-duration;
 
   @include size($element-size);
 
   &:checked {
 
     &::after {
-      content: '\00AC';
       position: absolute;
       top: 5%;
       left: 27%;
-      line-height: 1;
       font-size: ($element-size * 0.8);
-      font-weight: bold;
+      line-height: 1;
       transform: rotate(130deg);
+      content: '\00AC';
+      font-weight: bold;
     }
   }
 
@@ -43,20 +39,20 @@ $transition-duration: 0.2s;
     box-shadow: none;
 
     &::before {
-      content: '';
       position: absolute;
       top: 0;
       left: 0;
       display: block;
+      border-radius: $border-radius;
       box-shadow: inset 4px 4px 6px $box-shadow-first-color, inset -6px -6px 6px $box-shadow-second-color;
-      border-radius: 40%;
-      transition: box-shadow $transition-duration $transition-duration;    
+      transition: box-shadow $transition-duration $transition-duration;
+      content: '';    
 
       @include size(100%);
     }
   }
 
   &:hover {
-    color: $element-hover-cover;
+    color: $hover-color;
   }
 }
