+++ src/components/dropdown/dropdown.scss
@@ -1,73 +1,66 @@
-$background-color: #2C2F33;
-$box-shadow: 5px 5px 15px #000000, -5px -5px 15px #3B4451;
-$hover-color: rgba(37, 206, 222, 0.5);
+@import '../../styles/mixin.scss';
+
+@import '../../styles/variables.scss';
+
+$box-shadow: 5px 5px 15px $box-shadow-first-color, -5px -5px 15px $box-shadow-second-color;
 
 .dropdown {
     width: 275px;
     margin: 20px auto;
-    font-weight: bold;
-    color: #ffffff;
-  }
+    font-weight: 800;
+    color: $main-color;
 
-.dropdown__title {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
+    &__title {
+      display: flex;
+      justify-content: space-between;
+      width: 100%;
+      padding: 20px 40px;
+      color: $main-color;
+      background-color: $background-color;
+      border: none;
+      border-radius: 125px;
+      box-shadow: $box-shadow;
+      outline: none;
   
-    padding: 20px 40px;
-  
-    background: $background-color;
-  
-    box-shadow: $box-shadow;
-    border-radius: 124.46px;
-
-    border: none;
-    outline: none;
-    color: #ffffff;
-
-    &:hover {
-        color: $hover-color;
+      &:hover {
+          color: $hover-color;
+      }
     }
-  }
-  
-  .dropdown__content {
-    width: 100%;
-    height: 30px;
-    display: flex;
-    flex-direction: column;
-    margin-top: calc(-100%);
-    padding: 20px 20px 40px;
-  
-    background: $background-color;
-  
-    box-shadow: $box-shadow;
-    border-radius: 24px;
 
-    transition: margin-top 1s linear;
+    &__content {
+      display: flex;
+      flex-direction: column;
+      margin-top: calc(-100%);
+      padding: 20px 20px 40px;
+      overflow: hidden;
+      background-color: $background-color;
+      border-radius: $container-border-radius;
+      box-shadow: $box-shadow;
+      transition: margin-top 1s linear;
 
-    overflow: hidden;
-  }
+      @include size (100%, 30px);
 
-  .dropdown__content--open {
-    margin-top: 10px;
-    height: auto;
-  }
-  
-  .dropdown__item {
-    width: 100%;
-    margin-top: 10px;
+      &--open {
+        height: auto;
+        margin-top: 10px;
+      }
+    }
 
-    cursor: pointer;
-    &-text {
+    &__item {
+      width: 100%;
+      margin-top: 10px;
+      cursor: pointer;
+
+      &-text {
         margin-left: 20px;
-    }
-
-    & hr {
+      }
+  
+      hr {
         margin-top: 0;
-    }
-
-    &:hover {
+      }
+  
+      &:hover {
         color: $hover-color;
+      }
     }
-
   }
