+++ src/components/greeting/Greeting.scss
@@ -1,25 +1,17 @@
-$main-font-size: 16px;
-
-@mixin flex($flex-direction, $align-item: center, $justify-content: center) {
-  display: flex;
-  justify-content: $justify-content;
-  align-items: $align-item;
-}
+@import '../../styles/mixin.scss';
 
 .greeting {
-
-  @include flex(row, center, space-around);
   width: 100%;
-  cursor: default;
   user-select: none;
 
+  @include flex(row, center, space-around);
+
   .greeting__text {
-    font-size: $main-font-size;
+    font-size: 1.6rem;
     line-height: 1.5em;
 
     h1 {
-      margin: 0;
-      font-size: ($main-font-size * 2.25);
+      font-size: 3.6rem;
     }
   }
 }
