+++ src/components/switcher/Switcher.scss
@@ -1,45 +1,45 @@
+@import '../../styles/mixin.scss';
+
+@import '../../styles/variables.scss';
+
 $element-size: 30px;
-$main-color: #2C2F33;
 $transition-duration: 0.5s;
-$box-shadow-standard-color: #000000;
 $box-shadow-standard-setup: 4px 4px 6px;
 
-@mixin size($width, $height: $width) {
-  width: $width;
-  height: $height;
-}
+.switcher--stylise {
 
-.switcher_stylise {
   @include size(($element-size * 2), $element-size);
 
-  & input[type="checkbox"] {
+  input[type="checkbox"] {
     position: relative;
-    @include size(100%);
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    background-color: $main-color;
-    border-radius: ($element-size * 0.83);
+    background-color: $background-color;
+    border-radius: $container-border-radius;
+    box-shadow: inset $box-shadow-standard-setup $box-shadow-first-color, inset -2px -2px 6px $box-shadow-second-color;
     outline: none;
-    box-shadow: inset $box-shadow-standard-setup $box-shadow-standard-color, inset -2px -2px 6px #3B4451;
     transition: $transition-duration;
     cursor: pointer;
+    -webkit-appearance: none;
+    -moz-appearance: none;
+
+    @include size(100%);
 
     &::before {
-      content: '';
       position: absolute;
       top: 0;
       left: 0;
       margin: ($element-size * 0.1);
-      @include size(($element-size * 0.8));
+      background-color: rgb(238, 238, 238);
       border-radius: 50%;
-      background-color: #EEEEEE;
-      box-shadow: $box-shadow-standard-setup $box-shadow-standard-color;
+      box-shadow: $box-shadow-standard-setup $box-shadow-first-color;
       transition: $transition-duration;
+      content: '';
+
+      @include size(($element-size * 0.8));
     }
   }
 
-  & input:checked[type="checkbox"] {
-    background-color: #25CEDE;
+  input:checked[type="checkbox"] {
+    background-color: $enabled-color;
 
     &::before {
       left: 50%;
