+++ src/components/tabs/tabContent/TabContent.scss
@@ -1,27 +1,29 @@
-.tab-content {
-  display       : flex;
-  flex-direction: column;
+@import '../../../styles/mixin.scss';
 
+@import '../../../styles/variables.scss';
+
+.tab-content {
   width: 100%;
+
+  @include flex(column, stretch);
 
   &__header {
     padding-bottom: 5px;
-    border-bottom : 1px solid #FFFFFF;
+    border-bottom : 1px solid $main-color;
   }
 }
 
 .row {
-  display        : flex;
-  justify-content: space-between;
+
+  @include flex(row, stretch, space-between);
 
   &__item {
     width      : 25%;
     margin-top : 20px;
-    font-weight: normal;
-    font-size  : 18px;
-    line-height: 21px;
-    color      : #FFFFFF;
     text-align : left;
+    color      : $main-color;
+
+    @include font(1.8rem, 1.2em, 400)
   }
 
   &__item:last-child {
@@ -32,7 +34,7 @@
     cursor: pointer;
 
     &:hover {
-      color: #25CEDE; 
+      color: $enabled-color; 
       opacity: 0.5;
     }
   }
