+++ src/components/tabs/tabs.scss
@@ -1,24 +1,22 @@
-$background-color: #2c2f33;
-$hover-color: rgba(37, 206, 222, 0.5);
-$active-color: #25cede;
-$text-color: white;
-$shadow-light: #3b4451;
-$shadow-color: black;
+@import '../../styles/mixin.scss';
+
+@import '../../styles/variables.scss';
+
 $padding: 40px;
-$radius: 25px;
 
 .tabs {
-  background: $background-color;
+  font-family: inherit;
+  background-color: $background-color;
 
   &__tab-content {
     position: relative;
     z-index: 1;
     min-height: 150px;
     padding: $padding;
-    color: $text-color;
-    background: $background-color;
-    border-radius: $radius;
-    box-shadow: 6px 6px 19px $shadow-color, -6px -6px 19px $shadow-light;
+    color: $main-color;
+    background-color: $background-color;
+    border-radius: $container-border-radius;
+    box-shadow: $container-box-shadow;
   }
 
   &__tab-list {
@@ -29,25 +27,26 @@ $radius: 25px;
     &__tab-list-item {
       position: relative;
       display: inline-block;
-      width: 170px;
-      height: 40px;
       margin: 0 2rem;
-      background: $background-color;
-      border-radius: $radius $radius 0px 0px;
-      box-shadow: 6px 0px 16px $shadow-color, -6px -6px 19px $shadow-light;
+      background-color: $background-color;
+      border-radius: $container-border-radius $container-border-radius 0px 0px;
+      box-shadow: $container-box-shadow;
       list-style: none;
 
+      @include size(170px, 40px);
+
       button {
-        width: 100%;
-        height: 100%;
         padding: 0.5rem 0.75rem;
-        font-size: 24px;
-        color: $text-color;
-        background: transparent;
+        font-size: 2.4rem;
+        font-family: inherit;
+        color: $main-color;
+        background-color: transparent;
         border: none;
-        border-radius: $radius $radius 0px 0px;
+        border-radius: $container-border-radius $container-border-radius 0px 0px;
         outline: none;
         cursor: pointer;
+
+        @include size(100%);
       }
 
       button:hover {
@@ -56,13 +55,13 @@ $radius: 25px;
 
       &--active {
         z-index: 2;
-        background: $background-color;
-        border-radius: $radius $radius 0px 0px;
-        box-shadow: -6px -6px 19px $shadow-light;
+        background-color: $background-color;
+        border-radius: $container-border-radius $container-border-radius 0px 0px;
+        box-shadow: -6px -6px 18px $box-shadow-second-color;
         border-width: 1px 1px 0 1px;
 
         button {
-          color: $active-color;
+          color: $enabled-color;
         }
 
         button:hover {
