+++ src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDom from 'react-dom';
 
-import App from './app';
+import App from './pages/app/App';
 import 'semantic-ui-css/semantic.min.css';
+import './styles/index.scss';
 
 ReactDom.render(
   <App />,
