+++ src/pages/app/App.scss
@@ -1,0 +1,14 @@
+@import '../../styles/variables.scss';
+
+@import '../../styles/mixin.scss';
+
+.app-wrapper {
+    display: flex;
+}
+
+.app-main {
+    flex: 1 90%;
+    background-color: $background-color;
+
+    @include flex(column, stretch, stretch);
+}
