+++ src/pages/loginPage/LoginPage.scss
@@ -1,102 +1,103 @@
+@import '../../styles/mixin.scss';
+
+@import '../../styles/variables.scss';
+
 .login-wrapper {
-    display: flex;
-    height: 100vh;
-    align-items: center !important;
-    justify-content: center;
+  display: flex;
+  justify-content: center;
+  align-items: center !important;
+  height: 100vh;
 }
 
 .login-column {
-    width: 500px !important;
+  width: 500px !important;
 }
 
 .login-header {
-    color: #25CEDE;
-    text-align: center;
-    font-size: 64px;
-    line-height: 76px;
-    font-weight: bold;
+  text-align: center;
+  color: $enabled-color;
+
+  @include font(6.4rem, 1.2em, 800);
 }
 
 .login-header .image {
-    position: absolute !important;
-    left: -3.51%;
-    top: -14.45%;
+  position: absolute !important;
+  top: -14.45%;
+  left: -3.51%;
 }
 
 .login-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #2C2F33 !important;
-    box-shadow: 6.22302px 6.22302px 18.6691px #000000, -6.22302px -6.22302px 18.6691px #3B4451;
-    border-radius: 25px;
+  background-color: $background-color !important;
+  border-radius: $container-border-radius;
+  box-shadow: $container-box-shadow;
+
+  @include flex(column, center, stretch);
 }
 
 .login-form .label {
-    width: 200px;
-    margin: 25px auto;
-    color: #fff;
-    font-weight: 500;
-    font-size: 19.9809px;
-    line-height: 24px;
+  width: 200px;
+  margin: 25px auto;
+  color: $main-color;
+
+  @include font(2rem, 1.2em, 500);
 }
 
 .login-form .field {
-    width: 90%;
+  width: 90%;
 }
 
 .ui.fluid input {
-    padding-top: 15px !important;
-    padding-bottom: 15px !important;
-    background: #EEEEEE !important;
-    box-shadow: inset 5.9893px 5.9893px 17.9679px rgba(136, 160, 183, 0.25), inset -5.9893px -5.9893px 17.9679px rgba(255, 255, 255, 0.5) !important;
-    border-radius: 18px !important;
+  padding-top: 15px !important;
+  padding-bottom: 15px !important;
+  background: $main-color !important;
+  border-radius: 18px !important;
+  box-shadow: inset 6px 6px 18px rgba(136, 160, 183, 0.25), inset -6px -6px 18px rgba(255, 255, 255, 0.5) !important;
 }
 
 .checkbox-wrapper label {
-    color: #fff !important;
-    font-weight: 600;
+  color: $main-color !important;
+  font-weight: 600;
 }
 
 .ui.checkbox input:checked~.box:after, .ui.checkbox input:checked~label:after {
-    opacity: 1;
-    color: #6435c9 !important;
+  color: rgb(100, 53, 201) !important;
+  opacity: 1;
 }
 
 .field {
 	display: flex;
-	color: #fff;
+	color: $main-color;
 }
 
 .checkbox-wrapper {
-    flex: 0 68%;
+  flex: 0 68%;
 	font-weight: 600;
 }
 
 .field a {
-	color: #fff !important;
+	color: $main-color !important;
 	font-weight: 600;
 }
 
 .login-button {
 	width: 80%;
 	padding: 20px !important;
-	background: #25CEDE !important;
-	box-shadow: 12.446px 12.446px 31.1151px rgba(0, 0, 0, 0.5), -12.446px -12.446px 31.1151px #3B4451 !important;
-	border-radius: 24.8921px !important;
-	color: #fff !important;
+	font-size: 2.2rem !important;
+  line-height: 1.2em !important;
+	color: $main-color !important;
+	background: $enabled-color !important;
+	border-radius: $container-border-radius !important;
+	box-shadow: 12px 12px 31px rgba(0, 0, 0, 0.5), -12px -12px 31px rgb(59, 68, 81) !important;
 	font-weight: 500;
-	font-size: 21.5615px !important;
-	line-height: 26px !important;
 }
 
 .form .field:last-child {
 	width: 175px;
-    margin: 15px !important;
+  margin: 15px !important;
 }
 
 span.login-signup {
-	color: #fff !important;
+	color: $main-color !important;
 	font-weight: 400 !important;
 }
 
@@ -105,25 +106,25 @@ span.login-signup {
 }
 
 a.login-signup {
-	color: #25CEDE !important;
+	color: $enabled-color !important;
 }
 
 
 .user-notification {
-    position: absolute;
-    top: 80px;
-    height: 5vh;
-    width: 50%;
-    background-color: #25CEDE;
-    line-height: 5vh;
-    text-align: center;
-    border-radius: 20px;
-    font-weight: 600;
-    font-variant: small-caps;
-    color: #fff;
+  position: absolute;
+  top: 80px;
+  line-height: 5vh;
+  text-align: center;
+  color: $main-color;
+  background-color: $enabled-color;
+  border-radius: 20px;
+  font-weight: 500;
+  font-variant: small-caps;
+
+  @include size(50%, 5vh);
 }
 
 .user-notification.error {
-    background-color: #de2576;
+  background-color: rgb(222, 37, 118);
 }
