+++ src/styles/index.scss
@@ -1,0 +1,34 @@
+@import url('https://fonts.googleapis.com/css2?family=Baloo+Bhai+2:wght@400;500;700&family=Roboto:wght@400;500;700&display=swap');
+
+@import './variables.scss';
+
+@import './mixin.scss';
+
+html {
+  font-size: 10px;
+  line-height: 1.2em;
+}
+
+body {
+  min-width: 320px;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: "Baloo Bhai 2", "Roboto", sans-serif;
+  color: $main-color;
+  background-color: $background-color;
+  outline: none;
+  cursor: default;
+
+  @include font(1.5rem, 1,2em, 400);
+}
+
+h1, h2, h3, p {
+  margin: 0;
+  font-family: inherit;
+}
+
+*::after,
+*::before {
+  box-sizing: inherit;
+}
