+++ src/styles/mixin.scss
@@ -1,0 +1,18 @@
+@mixin size($width, $height: $width) {
+  width: $width;
+  height: $height;
+}
+
+@mixin flex($flex-direction, $align-items: center, $justify-content: center) {
+  display: flex;
+  flex-direction: $flex-direction;
+  justify-content: $justify-content;
+  align-items: $align-items;
+}
+
+@mixin font ($size, $height, $weight, $style: normal) {
+  font-style: $style;
+  font-size: $size;
+  line-height: $height;
+  font-weight: $weight;
+}
